backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
websockets==12.0

# GraphQL
strawberry-graphql[fastapi]==0.219.2

# Database
pymongo==4.6.1
motor==3.3.2

# AWS S3
boto3==1.34.34

# AI/ML
openai>=1.0.0
faiss-cpu==1.9.0
numpy==1.26.3
langchain==0.1.0
langchain-text-splitters==0.0.1

# Text Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Environment & Config
python-dotenv==1.0.1
pydantic==2.5.3

